# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticação",
        "description": "Documentação da API de autenticação",
        "contact": {
            "email": "gabriel.gomes@outlook.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Listar Usuários",
                "description": "Retorna uma lista de todos os usuários.",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Criar Usuário",
                "description": "Cria um novo usuário.",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "cpf",
                                    "nivel_acesso"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nome Usuário"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@email.com"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "nivel_acesso": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senhaSegura123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário criado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de Validação"
                                        },
                                        "erros": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Este email já foi cadastrado"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Atualizar Usuário",
                "description": "Atualiza as informações de um usuário.",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nome Usuário"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@email.com"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "nivel_acesso": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário atualizado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Deletar Usuário",
                "description": "Remove um usuário do sistema.",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Usuário",
                "required": [
                    "name",
                    "email",
                    "cpf",
                    "nivel_acesso"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nome do Usuário"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "usuario@email.com"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "123.456.789-00"
                    },
                    "nivel_acesso": {
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-10-16T15:51:05.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-10-16T15:51:05.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Insira o token no formato (Bearer \"token\")",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}